Added hook PortfolioAggregation/ViewFormPanel/useEditableView
