+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioAggregation/ViewFormPanel/useViewFormPanel.ts
@@ -1,30 +1,17 @@
 import { SPDataSourceItem } from 'pp365-shared-library'
-import { useContext, useEffect, useState } from 'react'
+import { useContext } from 'react'
 import { PortfolioAggregationContext } from '../context'
 import { TOGGLE_VIEW_FORM_PANEL } from '../reducer'
-
+import { useEditableView } from './useEditableView'
 
 /**
  * Hook for managing the logic of the `ViewFormPanel` component for the `PortfolioAggregation` component.
- * 
+ *
  * @returns An object containing functions and state variables for the `ViewFormPanel` component.
  */
 export function useViewFormPanel() {
   const context = useContext(PortfolioAggregationContext)
-  const [view, $setView] = useState(new Map<string, any>())
-  const isEditing = !!context.state.viewForm.view
-
-  useEffect(() => {
-    if (isEditing) {
-      $setView(new Map([
-        ['title', context.state.viewForm.view.title],
-        ['searchQuery', context.state.viewForm.view.searchQuery],
-        ['iconName', context.state.viewForm.view.iconName],
-      ]))
-    } else {
-      $setView(new Map())
-    }
-  }, [context.state.viewForm, context.state.currentView])
+  const { view, setView, isEditing } = useEditableView()
 
   /**
    * Dismisses the form panel by dispatching the `TOGGLE_VIEW_FORM_PANEL` action.
@@ -65,20 +52,6 @@ export function useViewFormPanel() {
   }
 
   /**
-   * Sets a view property.
-   *
-   * @param key Key of the view property to set
-   * @param value Value of the view property to set
-   */
-  const setView = (key: string, value: any) => {
-    $setView((prev) => {
-      const newView = new Map(prev)
-      newView.set(key, value)
-      return newView
-    })
-  }
-
-  /**
    * Determines whether the save button should be disabled based on the length of the view title and search query.
    */
   const isSaveDisabled = view.get('title')?.length < 2 || view.get('searchQuery')?.length < 51
@@ -88,6 +61,6 @@ export function useViewFormPanel() {
     isEditing: false,
     onDismiss,
     view,
-    setView,
+    setView
   } as const
 }
